orders/requirements.txt
aniso8601==9.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==7.1.2
Faker==14.0.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT==0.2.0
Flask-JWT-Extended==4.2.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==2.0.2
idna==3.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
psycopg2-binary==2.9.6
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.15
urllib3==1.26.15
Werkzeug==1.0.1
